refactor: moved root for readability and testability
