+++ cmd/serve.go
@@ -8,6 +8,7 @@
 package cmd
 
 import (
+	"github.com/robertwtucker/document-host/cmd/root"
 	"github.com/robertwtucker/document-host/internal/api"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
@@ -20,8 +21,8 @@ var serveCmd = &cobra.Command{
 	Long: `Starts the HTTP(S) server on the configured port and exposes the API endpoints
 `,
 	Run: func(cmd *cobra.Command, args []string) {
-		log.WithField("server", RootApp.Config.Server).Debug("starting API server")
-		app, err := api.NewApp(RootApp.Config)
+		log.WithField("server", root.Config.Server).Debug("starting API server")
+		app, err := api.NewApp(root.Config)
 		if err != nil {
 			return
 		}
@@ -30,5 +31,5 @@ var serveCmd = &cobra.Command{
 }
 
 func init() {
-	rootCmd.AddCommand(serveCmd)
+	root.Cmd().AddCommand(serveCmd)
 }
